Time: 0 ms (100%), Space: 8.7 MB (14.77%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -5,7 +5,7 @@ public:
         int high = x;
         int ans = 0;
         while(low<=high){
-            int mid = low + (high-low)/2;
+            long long mid = low + (high-low)/2;
             if (mid * mid <= x){
                 ans = mid;
                 low = mid +1;
